Further adaption for controlled movement

+++ include/game/elements/flight_track.hpp
@@ -25,7 +25,7 @@ public:
 
     void clear();
 
-    const core::graphics::Vector* moveToNextPoint(double f_distance);
+    const core::graphics::Vector* moveToNextPoint(float f_distance);
     const core::graphics::Vector* getPoint(int32_t f_index);
 
     void onMouseButtonPressed(const core::ui::MouseEventArgs& f_eventArgs) override;
